refactor: move EconomyEntry to economy/domain.go
