+++ economy/domain.go
@@ -1,9 +1,9 @@
-package models
+package economy
 
 import "github.com/google/uuid"
 
 type EconomyEntry struct {
-	UUID  uuid.UUID // Player’s UUID
+	UUID  uuid.UUID // Player's UUID
 	Name  string    // Display name
 	Money float64   // Balance
 }
